+++ BUILD
@@ -669,7 +669,10 @@ grpc_cc_library(
         "nofixdeps",
     ],
     visibility = ["@grpc:public"],
-    deps = ["grpc_public_hdrs"],
+    deps = [
+        "gpr_atm",
+        "grpc_public_hdrs",
+    ],
 )
 
 grpc_cc_library(
@@ -701,6 +704,7 @@ grpc_cc_library(
         "@grpc:public",
     ],
     deps = [
+        "gpr_atm",
         "grpc++_base",
         "slice",
     ],
@@ -7395,6 +7399,7 @@ grpc_cc_library(
         "env",
         "error",
         "gpr",
+        "gpr_atm",
         "gpr_manual_constructor",
         "grpc",
         "grpc++_codegen_proto",
@@ -7451,6 +7456,7 @@ grpc_cc_library(
         "channel_init",
         "config",
         "gpr",
+        "gpr_atm",
         "gpr_manual_constructor",
         "grpc_base",
         "grpc_health_upb",
