+++ include/grpcpp/alarm.h
@@ -24,7 +24,7 @@
 #include <functional>
 
 #include <grpc/grpc.h>
-#include <grpcpp/impl/codegen/completion_queue.h>
+#include <grpcpp/completion_queue.h>
 #include <grpcpp/impl/codegen/completion_queue_tag.h>
 #include <grpcpp/impl/codegen/grpc_library.h>
 #include <grpcpp/impl/codegen/time.h>
