+++ include/grpcpp/channel.h
@@ -22,9 +22,9 @@
 #include <memory>
 
 #include <grpc/grpc.h>
+#include <grpcpp/completion_queue.h>
 #include <grpcpp/impl/call.h>
 #include <grpcpp/impl/codegen/channel_interface.h>
-#include <grpcpp/impl/codegen/completion_queue.h>
 #include <grpcpp/impl/codegen/grpc_library.h>
 #include <grpcpp/impl/codegen/sync.h>
 #include <grpcpp/support/client_interceptor.h>
