+++ src/compiler/cpp_generator.cc
@@ -142,7 +142,7 @@ std::string GetHeaderIncludes(grpc_generator::File* file,
         "grpcpp/support/async_unary_call.h",
         "grpcpp/support/client_callback.h",
         "grpcpp/client_context.h",
-        "grpcpp/impl/codegen/completion_queue.h",
+        "grpcpp/completion_queue.h",
         "grpcpp/support/message_allocator.h",
         "grpcpp/support/method_handler.h",
         "grpcpp/impl/codegen/proto_utils.h",
