+++ test/cpp/codegen/compiler_test_golden
@@ -32,7 +32,7 @@
 #include <grpcpp/support/async_unary_call.h>
 #include <grpcpp/support/client_callback.h>
 #include <grpcpp/client_context.h>
-#include <grpcpp/impl/codegen/completion_queue.h>
+#include <grpcpp/completion_queue.h>
 #include <grpcpp/support/message_allocator.h>
 #include <grpcpp/support/method_handler.h>
 #include <grpcpp/impl/codegen/proto_utils.h>
